Runtime: Rename from interpreter
