+++ include/core/runtime/vm_register.hpp
@@ -3,13 +3,13 @@
 //
 #pragma once
 
-#include <core/interpreter/vm_register_common.hpp>
-#include <core/interpreter/vm_register_visitor.hpp>
+#include <core/runtime/vm_register_common.hpp>
+#include <core/runtime/vm_register_visitor.hpp>
 #include <core/exceptions.hpp>
 
 namespace dvm {
     namespace core {
-        namespace interpreter {
+        namespace runtime {
             namespace registers {
                 class vm_context;
