+++ include/core/runtime/vm_register_common.hpp
@@ -8,7 +8,7 @@
 
 namespace dvm {
     namespace core {
-        namespace interpreter {
+        namespace runtime {
             namespace registers {
                 /**
                  * 寄存器ID，用于标识寄存器
