--- src/core/interpreter/interpreter.cpp
@@ -1,5 +1,0 @@
-//
-// Created by kiva on 2017/8/28.
-//
-
-#include "core/interpreter/interpreter.hpp"
