+++ src/core/runtime/vm_context.cpp
@@ -2,11 +2,11 @@
 // Created by kiva on 2017/9/6.
 //
 
-#include <core/interpreter/vm_context.hpp>
+#include <core/runtime/vm_context.hpp>
 
 namespace dvm {
     namespace core {
-        namespace interpreter {
+        namespace runtime {
             vm_context::vm_context(UInt64 stack_size, UInt64 heap_size) {
             }
         }
