+++ src/core/runtime/vm_register_visitor.cpp
@@ -2,11 +2,11 @@
 // Created by kiva on 2017/9/3.
 //
 
-#include <core/interpreter/vm_register_visitor.hpp>
+#include <core/runtime/vm_register_visitor.hpp>
 
 namespace dvm {
     namespace core {
-        namespace interpreter {
+        namespace runtime {
             namespace registers {
                 register_visitor::register_visitor(VMRegister *register_ptr)
                     :register_ptr(register_ptr) {
